refactor a bit and add comments
